Move components into components/PAGE_NAME/
